requirements.txt
# Frigate Plus Reviewer (FPR) - Requirements

# Core dependencies
pyyaml>=6.0
requests>=2.31.0
Pillow>=10.0.0

# Vision model clients
google-generativeai>=0.3.0
openai>=1.0.0

# Optional testing
pytest>=7.4.0
pytest-cov>=4.1.0
requests-mock>=1.11.0
